+++ project/manager_frontend/views/monitor.py
@@ -1,8 +1,18 @@
 """
 Monitoring views
 """
+import os
+
 from django.conf import settings
 from django.views.generic import TemplateView
+
+# TODO: move to settings
+RECALBOX_THERMAL_BASEDIR = '/sys/class/thermal'
+# Assume 'thermal_zone0' is the CPU thermal infos on Raspberry
+RECALBOX_THERMAL_DEVICE_CPU_DIR = 'thermal_zone0'
+RECALBOX_THERMAL_CURRENT_TEMP_FILE = 'temp'
+RECALBOX_THERMAL_MAXTEMP_LIMIT_FILE = 'trip_point_0_temp'
+
 
 # Compatibility support for Recalbox versions from 3.2.x to 3.3.x
 # (psutil package only available since 3.3.0 beta 5)
@@ -18,6 +28,29 @@ else:
         """
         psutil_available = True
         mining_cpu_interval = settings.RECALBOX_PSUTIL_CPU_INTERVAL
+        
+        def get_thermal_infos(self, device_dir):
+            """
+            Watch for thermal infos using ACPI thermal API:
+            
+            http://lwn.net/Articles/268958/
+            
+            This is a naive implementation for the Raspberry2 which have only 
+            one "trip_point" that is the "hot" type (should be critical)
+            """
+            critical_limit = float(open(os.path.join(device_dir, RECALBOX_THERMAL_MAXTEMP_LIMIT_FILE), 'r').read().strip())
+            current_temp = float(open(os.path.join(device_dir, RECALBOX_THERMAL_CURRENT_TEMP_FILE), 'r').read().strip())
+            
+            # Divide per 1000 to have temperature in degrees Celsius
+            current_temp = current_temp/1000
+            critical_limit = critical_limit/1000
+            percent_usage = (current_temp/critical_limit)*100
+            
+            return {
+                'max': round(critical_limit, 2),
+                'current': round(current_temp, 2),
+                'percent_usage': round(percent_usage, 2),
+            }
         
         def get_cpu_infos(self):
             return {
@@ -66,5 +99,6 @@ class MonitoringView(RecalboxSystemInfosMixin, TemplateView):
                 'cpu_infos': self.get_cpu_infos(),
                 'memory_infos': self.get_memory_infos(),
                 'filesystem_infos': self.get_filesystem_infos(),
+                'cpu_thermal_infos': self.get_thermal_infos(os.path.join(RECALBOX_THERMAL_BASEDIR, RECALBOX_THERMAL_DEVICE_CPU_DIR)),
             })
         return context